update filter for newest version of CM, and update cmwiki cache.
NOTE WELL: the CM wiki is *really* crap for batter notes, and it shows in this updated cache as well.  For example, the LG G4 most certainly has a removable battery.  However, it doesn't show up in the current filter results because the CM wiki has zero mention of this.